zil: reorganise zil_lwb_flush_vdevs_done() locking
Wasabi have reported the following shape of crash:

    Aug 24 19:44:09 C2-R106-U40 kernel: general protection fault, probably for non-canonical address 0xa0bba1ad27732c24: 0000 [#1] SMP NOPTI
    ...
    Aug 24 19:44:09 C2-R106-U40 kernel: RIP: 0010:__mutex_lock.isra.0+0xc2/0x470
    ...
    Aug 24 19:44:09 C2-R106-U40 kernel: Call Trace:
    Aug 24 19:44:09 C2-R106-U40 kernel:  <TASK>
    Aug 24 19:44:09 C2-R106-U40 kernel:  ? __wake_up_common_lock+0x8a/0xc0
    Aug 24 19:44:09 C2-R106-U40 kernel:  __mutex_lock_slowpath+0x13/0x20
    Aug 24 19:44:09 C2-R106-U40 kernel:  mutex_lock+0x36/0x40
    Aug 24 19:44:09 C2-R106-U40 kernel:  zil_lwb_flush_vdevs_done+0x1bc/0x380 [zfs]
    Aug 24 19:44:09 C2-R106-U40 kernel:  zio_done+0x312/0x15f0 [zfs]
    Aug 24 19:44:09 C2-R106-U40 kernel:  zio_execute+0x92/0xf0 [zfs]
    Aug 24 19:44:09 C2-R106-U40 kernel:  taskq_thread+0x2bf/0x4d0 [spl]
    Aug 24 19:44:09 C2-R106-U40 kernel:  ? wake_up_q+0x90/0x90
    Aug 24 19:44:09 C2-R106-U40 kernel:  ? zio_taskq_member.isra.0.constprop.0+0xa0/0xa0 [zfs]
    Aug 24 19:44:09 C2-R106-U40 kernel:  ? task_done+0xb0/0xb0 [spl]
    Aug 24 19:44:09 C2-R106-U40 kernel:  kthread+0x127/0x150
    Aug 24 19:44:09 C2-R106-U40 kernel:  ? set_kthread_struct+0x50/0x50
    Aug 24 19:44:09 C2-R106-U40 kernel:  ret_from_fork+0x1f/0x30
    Aug 24 19:44:09 C2-R106-U40 kernel:  </TASK>

This is a weird one. Its difficult to correlate exactly, but as best I
can tell, this is trying to take either zl_lock or zl_issuer_lock near
the top of zil_lwb_flush_vdevs_done(). Since they're on the
dataset-global zilog_t, and has the same lifetime as zl_lock, its not
entirely clear how this could be a use-after-free or a corruption, yet
that's what it looks like.

No error or disk failure was reported in both the occurrences we've
seen, so it seems unlikely that this was an error, and so io_error
should be zero and so zl_issuer_lock never taken. But, I don't see any
way that zl_lock could really be involved.

I started thinking about the MUTEX_HELD() check when the ZIL is failed
and we need to release zl_issuer_lock, and how its not entirely
symmetrrical with the error check decision to acquire the lock. I can't
think of any path where we could be trying to release zl_issuer_lock
when we don't hold it, but there is an enormously implausible way if we
were to leak zl_issuer_lock on the same taskq thread that eventually
calls zil_lwb_flush_vdevs_done(), and then and up falsely acquiring or
releasing it. I don't think that's even in play, but I did see that it
was silly to take zl_issuer_lock if we're not actually going to fail the
ZIL, and instead we can just trade locks if we need to.

I still don't see how this solves the actual problem though, sigh.

(cherry picked from commit dda0e91eadc47703f4e3910e2efc1738139ef1a7)
Signed-off-by: Allan Jude <allan@klarasystems.com>