zil_commit_waiter: handle cv_wait() returning before waiter has completed
In rare situations, it's possible for a thread waiting in cv_wait() to
be woken by the kernel before the condition has been met. If this
happens, we return into zil_commit_impl() and the waiter can be freed
while still on the lwb_waiters list, corrupting the list.

To fix, we wrap the wait logic in a loop. If the thread is woken and
zcw_done is not set, we go back to sleep waiting for it to complete. The
loop now cannot exit until and unless the waiter is completed and
removed from the lwb_waiters list, making it safe to free.

Signed-off-by: Rob Norris <rob.norris@klarasystems.com>